# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeleShop API — crypto marketplace for Telegram Mini Apps
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramAuthInput {
  initData: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  role: string;
  /** @nullable */
  referralCode?: string | null;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface UserUpdate {
  username?: string;
  firstName?: string;
  lastName?: string;
}

export interface Shop {
  id: number;
  ownerId: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  plan: string;
  status: string;
  rating?: string;
  totalSales: number;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  telegramChannel?: string | null;
  createdAt: string;
}

export interface ShopList {
  items: Shop[];
  total: number;
}

export type ShopInputPlan = typeof ShopInputPlan[keyof typeof ShopInputPlan];


export const ShopInputPlan = {
  starter: 'starter',
  growth: 'growth',
  pro: 'pro',
} as const;

export interface ShopInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  slug: string;
  description?: string;
  plan: ShopInputPlan;
  walletAddress?: string;
  telegramChannel?: string;
}

export interface ShopUpdate {
  name?: string;
  description?: string;
  logoUrl?: string;
  bannerUrl?: string;
  walletAddress?: string;
  telegramChannel?: string;
  status?: string;
}

export interface Product {
  id: number;
  shopId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  currency: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  stock: number;
  status: string;
  /** @nullable */
  flashSalePrice?: string | null;
  /** @nullable */
  flashSaleEnd?: string | null;
  createdAt: string;
}

export interface ProductList {
  items: Product[];
  total: number;
}

export type ProductInputCurrency = typeof ProductInputCurrency[keyof typeof ProductInputCurrency];


export const ProductInputCurrency = {
  TON: 'TON',
  USDT: 'USDT',
} as const;

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  price: string;
  currency: ProductInputCurrency;
  imageUrl?: string;
  category?: string;
  stock: number;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: string;
  stock?: number;
  status?: string;
  imageUrl?: string;
  category?: string;
  flashSalePrice?: string;
  flashSaleEnd?: string;
}

export interface Order {
  id: number;
  buyerId: number;
  shopId: number;
  productId: number;
  quantity: number;
  totalAmount: string;
  currency: string;
  status: string;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  escrowAddress?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  voucherCode?: string | null;
  /** @nullable */
  discountAmount?: string | null;
  createdAt: string;
}

export interface OrderList {
  items: Order[];
  total: number;
}

export interface OrderInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
  deliveryAddress?: string;
  note?: string;
  voucherCode?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  confirmed: 'confirmed',
  shipped: 'shipped',
  completed: 'completed',
  disputed: 'disputed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  txHash?: string;
}

export interface Voucher {
  id: number;
  /** @nullable */
  shopId?: number | null;
  code: string;
  discountType: string;
  discountValue: string;
  /** @nullable */
  minOrderAmount?: string | null;
  /** @nullable */
  maxUses?: number | null;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface VoucherList {
  items: Voucher[];
}

export type VoucherInputDiscountType = typeof VoucherInputDiscountType[keyof typeof VoucherInputDiscountType];


export const VoucherInputDiscountType = {
  percent: 'percent',
  fixed: 'fixed',
} as const;

export interface VoucherInput {
  /** @minLength 3 */
  code: string;
  discountType: VoucherInputDiscountType;
  discountValue: string;
  minOrderAmount?: string;
  maxUses?: number;
  expiresAt?: string;
}

export interface VoucherValidateInput {
  code: string;
  orderAmount: string;
  shopId?: number;
}

export interface VoucherValidateResult {
  valid: boolean;
  voucher: Voucher;
  discountAmount?: string;
  finalAmount?: string;
}

export interface Affiliate {
  id: number;
  userId: number;
  referralCode: string;
  tier: string;
  totalEarned: string;
  pendingPayout: string;
  totalReferrals: number;
  status: string;
  createdAt: string;
}

export type ListShopsParams = {
page?: number;
limit?: number;
category?: string;
search?: string;
};

export type ListProductsParams = {
page?: number;
limit?: number;
search?: string;
category?: string;
flashSale?: boolean;
shopId?: number;
};

export type ListOrdersParams = {
role?: ListOrdersRole;
status?: string;
page?: number;
limit?: number;
};

export type ListOrdersRole = typeof ListOrdersRole[keyof typeof ListOrdersRole];


export const ListOrdersRole = {
  buyer: 'buyer',
  seller: 'seller',
} as const;