# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: TeleShop API — crypto marketplace for Telegram Mini Apps
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/telegram:
    post:
      operationId: authTelegram
      tags: [auth]
      summary: Authenticate via Telegram initData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid initData
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products across all shops (marketplace feed)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: flashSale
          schema: { type: boolean }
        - in: query
          name: shopId
          schema: { type: integer }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product in current user's shop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user (buyer or seller)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [buyer, seller]
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (confirm, ship, complete, dispute)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /affiliates/me:
    get:
      operationId: getMyAffiliate
      tags: [affiliates]
      summary: Get current user's affiliate profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Affiliate profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affiliate"
        "404":
          description: Not enrolled
    post:
      operationId: enrollAffiliate
      tags: [affiliates]
      summary: Enroll in affiliate program
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affiliate"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TelegramAuthInput:
      type: object
      required: [initData]
      properties:
        initData:
          type: string
        referralCode:
          type: ["string", "null"]
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, telegramId, role, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        role:
          type: string
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, shopId, name, price, currency, stock, status, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        stock:
          type: integer
        status:
          type: string
        flashSalePrice:
          type: ["string", "null"]
        flashSaleEnd:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ProductList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
    ProductInput:
      type: object
      required: [name, price, currency, stock]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: string
        currency:
          type: string
          enum: [TON, USDT]
        imageUrl:
          type: string
        category:
          type: string
        stock:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        stock:
          type: integer
        status:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        flashSalePrice:
          type: string
        flashSaleEnd:
          type: string
    Order:
      type: object
      required: [id, buyerId, shopId, productId, quantity, totalAmount, currency, status, createdAt]
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        shopId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        totalAmount:
          type: string
        currency:
          type: string
        status:
          type: string
        txHash:
          type: ["string", "null"]
        escrowAddress:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        voucherCode:
          type: ["string", "null"]
        discountAmount:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OrderList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
    OrderInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        deliveryAddress:
          type: string
        note:
          type: string
        voucherCode:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, shipped, completed, disputed, cancelled]
        txHash:
          type: string
...
    Affiliate:
      type: object
      required: [id, userId, referralCode, tier, totalEarned, pendingPayout, totalReferrals, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        referralCode:
          type: string
        tier:
          type: string
        totalEarned:
          type: string
        pendingPayout:
          type: string
        totalReferrals:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time