# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeleShop API — crypto marketplace for Telegram Mini Apps
 * OpenAPI spec version: 1.0.0
 */

export interface Order {
  id: number;
  buyerId: number;
  shopId: number;
  productId: number;
  quantity: number;
  totalAmount: string;
  currency: string;
  status: string;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  escrowAddress?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  voucherCode?: string | null;
  /** @nullable */
  discountAmount?: string | null;
  createdAt: Date;
}