# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeleShop API — crypto marketplace for Telegram Mini Apps
 * OpenAPI spec version: 1.0.0
 */

export interface Product {
  id: number;
  shopId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  currency: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  stock: number;
  status: string;
  /** @nullable */
  flashSalePrice?: string | null;
  /** @nullable */
  flashSaleEnd?: Date | null;
  createdAt: Date;
}