# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeleShop API — crypto marketplace for Telegram Mini Apps
 * OpenAPI spec version: 1.0.0
 */
import type { ProductInputCurrency } from './productInputCurrency';

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  price: string;
  currency: ProductInputCurrency;
  imageUrl?: string;
  category?: string;
  stock: number;
}