# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeleShop API — crypto marketplace for Telegram Mini Apps
 * OpenAPI spec version: 1.0.0
 */

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  role: string;
  /** @nullable */
  referralCode?: string | null;
  createdAt: Date;
}